stuff:
- set mouse cursor when dragging plot
- move PlotPanner class into separate file
